requirements.txt
amqp==5.0.6
apturl==0.5.2
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
boto3==1.20.8
botocore==1.23.9
Brlapi==0.7.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
cryptography==35.0.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.9
django-allauth==0.45.0
django-braces==1.15.0
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
duplicity==0.8.12.0
dynaconf==3.1.7
entrypoints==0.3
fasteners==0.14.1
filelock==3.4.0
future==0.18.2
httplib2==0.14.0
hvac==0.11.2
idna==3.3
install==1.3.4
Jinja2==3.0.3
jmespath==0.10.0
jwcrypto==1.0
keyring==18.0.1
knox==0.1.14
kombu==5.2.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
loguru==0.5.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.1
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.9.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
s3transfer==0.5.0
SecretStorage==2.3.1
shortuuid==1.0.8
simplejson==3.16.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
systemd-python==234
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.7
validators==0.18.2
vine==5.0.0
virtualenv==20.10.0
wadllib==1.3.3
wcwidth==0.2.5
wrapt==1.13.3
xkit==0.0.0
